requirements.txt
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
google-auth==2.33.0
google-auth-oauthlib==1.2.2
greenlet==3.2.4
gspread==6.1.4
idna==3.10
numpy==2.0.2
oauthlib==3.3.1
pandas==2.2.2
playwright==1.55.0
playwright-stealth==1.0.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyee==13.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
